+++ src/components/search/Search.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react';
 
 import SpotifyWebApi from '../../spotify';
